requirements.txt
requests
PyYAML
pytest
jinja2
